+++ hss_okta_flutter/lib/src/web/stubs/hss_okta_js_mobile_stub.dart
@@ -51,6 +51,8 @@ class OktaAuth {
   void setOriginalUri(String uri) => throw UnimplementedError();
   void removeOriginalUri() => throw UnimplementedError();
   void getOriginalUri() => throw UnimplementedError();
+
+  void setHeaders(Object headers) => throw UnimplementedError();
 }
 
 class OktaConfig {