+++ hss_okta_flutter/CHANGELOG.md
@@ -52,3 +52,5 @@
 * Added revoke RefreshToken
 * Added Alias of getUserInfo
 * Added removeOriginalUri,getOriginalUri, and setOriginalUri to OktaAuth
+* Added handleRedirect for OktaAuth
+* Added setHeaders to OktaAuth